Update PDFmaker.py
Removed the boilerplate tkinter code from the last commit, it was copied from somewhere but it didn't do what I wanted. I couldn't get the name of the chosen file to display correctly, the font wouldn't change. etc. Tried messing with it but ultimately went with the "just start over" approach.